+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Fixed kwargs error when assigning `vertexcolor`.
+* Fixed a bug of mis-matching colors and positions length for `GeometryObject` buffer data.
 
 ### Removed